Add GitHub workflow
In order to make sure that any PR on GitHub keeps the publishing
process intact, GitHub workflows/actions are there to save the day.

This commit introduces two jobs:

== Elisp-checks
This check uses the leotaku/elisp-check action to ensure that the file
can get loaded by emacs and that the documentation follows best
practices. Note that there are a lot of warnings currently on
`checkdoc`. We could set `warnings_as_errors` to `true`, but that is
probably better *after* we made sure that those errors are gone first.

== Check-publish
This is a sanity check to ensure that ./local-build.sh finishes
without any errors.